+++ toothpick-compiler/src/test/java/toothpick/compiler/factory/RelaxedFactoryForClassContainingMethodsTest.java
@@ -46,7 +46,7 @@ public class RelaxedFactoryForClassContainingMethodsTest extends BaseFactoryTest
         "    return false;", //
         "  }", //
         "  @Override", //
-        "  public boolean hasScopeInstancesAnnotation() {", //
+        "  public boolean hasProvidesSingletonInScopeAnnotation() {", //
         "    return false;", //
         "  }", //
         "}" //