+++ toothpick/src/main/java/toothpick/config/Module.java
@@ -2,7 +2,7 @@ package toothpick.config;
 
 import java.util.HashSet;
 import java.util.Set;
-import toothpick.ScopeInstances;
+import toothpick.ProvidesSingletonInScope;
 
 /**
  * Defines a set of bindings, {@code Module} instances are installed in a scope.
@@ -19,7 +19,7 @@ import toothpick.ScopeInstances;
  * <dd>binds an interface or class to a provider class. <br/>
  * If class FooProvider is annotated with the {@link javax.inject.Singleton} annotation, the same instance of the provider is reused for each
  * injection. <br/>
- * If class FooProvider is annotated with the {@link ScopeInstances} annotation, the same instance of Foo is reused for each
+ * If class FooProvider is annotated with the {@link ProvidesSingletonInScope} annotation, the same instance of Foo is reused for each
  * injection. <br/>
  * Otherwise, a new instance of class Foo will be produced by a new instance of FooProvider for each injection.</dd>
  * <dt><tt>bind(IFoo.class).toProvider(new FooProvider())</tt></dt>
@@ -70,7 +70,7 @@ import toothpick.ScopeInstances;
  * <p>
  * <b>{@link javax.inject.Provider} &amp; recycling of instances :</b><br>
  * If providers are bound by class, the annotations
- * {@link javax.inject.Singleton} and {@link ScopeInstances} can be used to precise how instances of the provider
+ * {@link javax.inject.Singleton} and {@link ProvidesSingletonInScope} can be used to precise how instances of the provider
  * and instances produced by this provider are recycled across multiple injections. <br/>
  * If providers are bound by instances, then the provider has to deal by itself with the recycling of the instance it produces.
  * </p>